packages/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pnlclaw-cli"
version = "0.1.0"
description = "PnLClaw command-line interface"
readme = { text = "CLI for PnLClaw Community", content-type = "text/plain" }
license = "AGPL-3.0-only"
requires-python = ">=3.11"
dependencies = [
  "click>=8.1",
  "pnlclaw-types",
  "pnlclaw-core",
  "pnlclaw-strategy",
  "pnlclaw-backtest",
  "pnlclaw-paper",
  "pnlclaw-risk",
  "pnlclaw-storage",
  "pnlclaw-security",
  "httpx>=0.27.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[project.scripts]
pnlclaw = "pnlclaw_cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["pnlclaw_cli"]
